ping-pong-app/requirements.txt
Flask==2.1.3
Werkzeug==2.0.0
psycopg2-binary==2.9.9